01_Smart_Student_Agent_Assistant/requirements.txt
openai-agents>=0.0.19
python-dotenv>=1.0.0

